requirements.txt
asgiref==3.2.7
beautifulsoup4==4.6.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.1.14
django-cors-headers==3.2.1
django-filter==2.2.0
gunicorn==20.0.4
hangul-romanize==0.1.0
idna==2.9
JPype1==0.7.3
lxml==4.5.0
mecab-python3==0.996.5
numpy==1.18.3
oauthlib==3.1.0
psycopg2==2.8.5
pybind11==2.5.0
PySocks==1.7.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
tweepy==3.8.0
urllib3==1.25.9
whitenoise==5.0.1
python-mecab-ko==1.0.12
